requirements.txt
beautifulsoup4==4.14.2
boto3==1.40.57
botocore==1.40.57
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
idna==3.11
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.5.2
lxml==6.0.2
lxml_html_clean==0.4.3
newspaper3k==0.2.8
nltk==3.9.2
pillow==12.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-file==3.0.1
s3transfer==0.14.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
